Complete the Projects Card
