+++ src/main.jsx
@@ -14,7 +14,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import App from './App';
 import Error from './pages/Error';
 import Home from './pages/home';
-import Projects from './pages/projects';
+import Projects from './pages/projects/projects';
 import Contact from './pages/contact/contact';
 import About from './pages/about/about';
