+++ src/pages/projects/projectheader.jsx
@@ -1,0 +1,16 @@
+/*******************************************************************
+ * Carleton Bootcamp - 2024
+ * Copyright 2024 Gustavo Miller
+ * License: free and unencumbered software
+ * Assignment # 20 - React Portfolio
+ * 
+ * Filename: projectheader.jsx
+ * Date : 1/14/2024 22:07:18 PM
+ *******************************************************************/
+export default function Contactheader() {
+     return (
+          <div className="text-center mt-5 ">
+               <h1 className="mb-5">Projects Developed</h1>
+               <p><strong>Application Developed</strong>. Throughout my career I have developed many solutions. Since I started the Carletong Coding Bootcamp I have developed the following application using new technologies. Some of the technology stack used include Node.js, Javascript, ExpressJS, MongoDB, MySql and much more.</p>
+          </div>
+     )};
